+++ src/prefect/tasks/__init__.py
@@ -8,4 +8,5 @@ import prefect.tasks.database
 import prefect.tasks.docker
 import prefect.tasks.github
 import prefect.tasks.notifications
+import prefect.tasks.secrets
 import prefect.tasks.shell