+++ src/prefect/tasks/snowflake/snowflake.py
@@ -98,7 +98,7 @@ class SnowflakeQuery(Task):
         try:
             with conn:
                 with conn.cursor() as cursor:
-                    executed = cursor.execute(query=query, params=data)
+                    executed = cursor.execute(query, params=data)
 
             conn.close()
             return executed